+++ catalogue/DriveConfig.cpp
@@ -58,6 +58,7 @@ void DriveConfig::setTapedConfiguration(const cta::tape::daemon::TapedConfigurat
   setConfigToDB(&config->useMaintenanceProcess, catalogue, tapeDriveName);
   setConfigToDB(&config->externalFreeDiskSpaceScript, catalogue, tapeDriveName);
   setConfigToDB(&config->tapeLoadTimeout, catalogue, tapeDriveName);
+  setConfigToDB(&config->wdGlobalLockAcqMaxSecs, catalogue, tapeDriveName);
 }
 
 void DriveConfig::checkConfigInDB(catalogue::Catalogue* catalogue, const std::string& tapeDriveName,