+++ cmdline/CtaAdminCmd.cpp
@@ -276,6 +276,16 @@ void CtaAdminCmd::send() const
       throw std::runtime_error("Configuration error: cta.endpoint missing from " + config_file.string());
    }
 
+   // Check drive timeout value
+   auto [driveTimeoutConfigExists, driveTimeoutVal] = config.getOptionValueInt("drive_timeout");
+   if(driveTimeoutConfigExists) {
+     if (driveTimeoutVal > 0) {
+       formattedText.setDriveTimeout(driveTimeoutVal);
+     } else {
+       throw std::runtime_error("Configuration error: cta.drive_timeout not a positive value in " + config_file.string());
+     }
+   }
+
    // If the server is down, we want an immediate failure. Set client retry to a single attempt.
    XrdSsiProviderClient->SetTimeout(XrdSsiProvider::connect_N, 1);