+++ objectstore/BackendRados.cpp
@@ -562,7 +562,7 @@ void BackendRados::lockNotify(const std::string& name, uint64_t timeout_us, Lock
     watcher.wait(watchTimeout);
     TIMESTAMPEDPRINT(lockType==LockType::Shared?"Post-wait (shared)":"Post-wait (exclusive)");
     if (timeout_us && (timeoutTimer.usecs() > (int64_t)timeout_us)) {
-      throw exception::Exception("In BackendRados::lockNotify(): timeout.");
+      throw exception::TimeoutException("In BackendRados::lockNotify(): timeout.");
     }
   }
   cta::exception::Errnum::throwOnReturnedErrno(-rc,
@@ -750,7 +750,7 @@ void BackendRados::lockBackoff(const std::string& name, uint64_t timeout_us, Loc
     }
     if (-EBUSY != rc) break;
     if (timeout_us && (timeoutTimer.usecs() > (int64_t)timeout_us)) {
-      throw exception::Exception(
+      throw exception::TimeoutException(
         "In BackendRados::lockBackoff(): timeout : timeout set = " + std::to_string(timeout_us) + " usec, time to lock the object : " +
         std::to_string(timeoutTimer.usecs()) + " usec, number of tries to lock = " + std::to_string(nbTriesToLock) + " object: " + name);
     }