+++ tapeserver/daemon/TapedConfiguration.hpp
@@ -134,6 +134,9 @@ struct TapedConfiguration {
   /// Time to wait after scheduling came up idle
   cta::SourcedParameter<time_t> wdIdleSessionTimer{
     "taped", "WatchdogIdleSessionTimer", 10, "Compile time default"};
+  /// Time to wait after which the tape server stops trying to get the next mount
+  cta::SourcedParameter<time_t> wdGlobalLockAcqMaxSecs{
+    "taped", "WatchdogGlobalLockAcqMaxSecs", 15 * 60, "Compile time default"};
   //----------------------------------------------------------------------------
   // The central storage access configuration
   //----------------------------------------------------------------------------